20211215-01

+++ backend/BEListEdit.js
@@ -5,8 +5,8 @@ exports.insert_main = function(req, res) {
   res.header("Content-Type", "application/json; charset=utf-8");
   var apiName = '[insert_main]';
   console.log(apiName+': start');
-  console.log('group_no:'+req.query.group_no+', name:'+req.query.name+', sort:'+req.query.sort);
-  exec_insert([req.query.group_no, req.query.name, req.query.sort]);
+  console.log('user_id:'+req.query.user_id+', group_no:'+req.query.group_no+', name:'+req.query.name+', sort:'+req.query.sort);
+  exec_insert([req.query.user_id, req.query.group_no, req.query.name, 'dcdcdc', req.query.sort]);
 
   function return_result(flg) {
     console.log('err: '+flg);
@@ -24,7 +24,7 @@ exports.insert_main = function(req, res) {
     var err_flg = false;
     var cnt = 0;
     const dbm = await getPostgresClient();
-    const sql = 'INSERT INTO TO_DO_LIST(GROUP_NO,NAME,SORT) VALUES($1, $2, $3)';
+    const sql = 'INSERT INTO TO_DO_LIST(USER_ID,GROUP_NO,NAME,COLOR,SORT) VALUES($1, $2, $3, $4, $5)';
     try {
       await dbm.begin();
       const result = await dbm.execute(sql, params);
@@ -46,8 +46,8 @@ exports.delete_main = function(req, res) {
   res.header("Content-Type", "application/json; charset=utf-8");
   var apiName = '[delete_main]';
   console.log(apiName+': start');
-  console.log('group_no:'+req.query.group_no);
-  exec_delete([req.query.group_no]);
+  console.log('user_id:'+req.query.user_id+', group_no:'+req.query.group_no);
+  exec_delete([req.query.group_no, req.query.user_id]);
 
   function return_result(flg) {
     console.log('err: '+flg);
@@ -65,9 +65,9 @@ exports.delete_main = function(req, res) {
     var err_flg = false;
     var cnt = 0;
     const dbm = await getPostgresClient();
-    const sql = 'DELETE FROM TO_DO_LIST WHERE GROUP_NO = $1';
-    const sql2 = 'DELETE FROM TO_DO_LIST_HED WHERE GROUP_NO = $1';
-    const sql3 = 'DELETE FROM TO_DO_LIST_DTL WHERE GROUP_NO = $1';
+    const sql = 'DELETE FROM TO_DO_LIST WHERE GROUP_NO = $1 AND USER_ID = $2';
+    const sql2 = 'DELETE FROM TO_DO_LIST_HED WHERE GROUP_NO = $1 AND USER_ID = $2';
+    const sql3 = 'DELETE FROM TO_DO_LIST_DTL WHERE GROUP_NO = $1 AND USER_ID = $2';
     try {
       await dbm.begin();
       const result = await dbm.execute(sql, params);
@@ -91,8 +91,8 @@ exports.update_main_name = function(req, res) {
   res.header("Content-Type", "application/json; charset=utf-8");
   var apiName = '[update_main_name]';
   console.log(apiName+': start');
-  console.log('group_no:'+req.query.group_no+', name:'+req.query.name);
-  update_main([req.query.group_no, req.query.name]);
+  console.log('user_id:'+req.query.user_id+', group_no:'+req.query.group_no+', name:'+req.query.name+', color:'+req.query.color);
+  update_main([req.query.group_no, req.query.user_id, req.query.name, req.query.color]);
 
   function return_result(flg) {
     console.log('err: '+flg);
@@ -110,7 +110,7 @@ exports.update_main_name = function(req, res) {
     var err_flg = false;
     var cnt = 0;
     const dbm = await getPostgresClient();
-    const sql = 'UPDATE TO_DO_LIST SET NAME = $2 WHERE GROUP_NO = $1';
+    const sql = 'UPDATE TO_DO_LIST SET NAME = $3, COLOR = $4 WHERE GROUP_NO = $1 AND USER_ID = $2';
     try {
       await dbm.begin();
       const result = await dbm.execute(sql, params);
@@ -149,16 +149,16 @@ exports.update_main_sort = function(req, res) {
   async function update_main(items) {
     var err_flg = false;
     var log_message = '';
-    var no;
+    var user_id;
     const dbm = await getPostgresClient();
-    const sql = 'UPDATE TO_DO_LIST SET SORT = $2 WHERE GROUP_NO = $1';
+    const sql = 'UPDATE TO_DO_LIST SET SORT = $3 WHERE GROUP_NO = $1 AND USER_ID = $2';
     try {
       await dbm.begin();
       items.map(async item => {
-        const params = [item.group_no, item.sort];
+        const params = [item.group_no, item.user_id, item.sort];
         await dbm.execute(sql, params);
         log_message += '('+item.group_no+','+item.sort+'),';
-        no = item.no;
+        user_id = item.user_id;
       });
       await dbm.commit();
     } catch (e) {
@@ -167,7 +167,7 @@ exports.update_main_sort = function(req, res) {
       err_flg = true;
     } finally {
       await dbm.release();
-      console.log('no: '+no);
+      console.log('user_id: '+user_id);
       console.log('group_no,sort: '+log_message);
       return_result(err_flg);
     }
